+++ app/src/main/java/kolomyichuk/runly/ui/MainScreen.kt
@@ -1,10 +1,9 @@
-package kolomyichuk.runly.navigation
+package kolomyichuk.runly.ui
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -13,22 +12,26 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import kolomyichuk.runly.ui.navigation.Screen
+import kolomyichuk.runly.ui.navigation.profileNavGraph
+import kolomyichuk.runly.ui.navigation.runNavGraph
 import kolomyichuk.runly.ui.components.BottomNavigationBar
 import kolomyichuk.runly.ui.screens.home.HomeScreen
-import kolomyichuk.runly.ui.viewmodel.HomeViewModel
+import kolomyichuk.runly.ui.screens.home.HomeViewModel
 
 @Composable
 fun MainScreen(
-    startScreen: MutableState<Screen>
+    currentScreen: Screen,
+    onCurrentScreenChange: () -> Unit
 ) {
     val navController = rememberNavController()
 
-    LaunchedEffect(startScreen.value) {
-        startScreen.value.let {
+    LaunchedEffect(currentScreen) {
+        currentScreen.let {
             navController.navigate(it) {
                 popUpTo(Screen.Home) { inclusive = false }
             }
-            startScreen.value = Screen.Home
+            onCurrentScreenChange()
         }
     }
