+++ app/src/main/java/kolomyichuk/runly/ui/screens/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package kolomyichuk.runly.ui.viewmodel
+package kolomyichuk.runly.ui.screens.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
