+++ app/src/main/java/kolomyichuk/runly/ui/screens/run/RunScreen.kt
@@ -35,11 +35,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import kolomyichuk.runly.R
 import kolomyichuk.runly.data.local.datastore.AppTheme
-import kolomyichuk.runly.navigation.Screen
 import kolomyichuk.runly.service.RunTrackingService
 import kolomyichuk.runly.ui.components.TopBarApp
-import kolomyichuk.runly.ui.viewmodel.RunViewModel
-import kolomyichuk.runly.ui.viewmodel.ThemeViewModel
+import kolomyichuk.runly.ui.screens.theme.ThemeViewModel
 
 @Composable
 fun RunScreen(
